fix: Prevent OpenSeadragon canvas from being selected on iOS devices (#491)
This commit addresses an issue where the OpenSeadragon canvas element could be inadvertently selected on iOS devices. By applying the user-select: none CSS property to the OpenSeadragon canvas, we ensure that users cannot select this element, thereby improving the overall user experience on iOS.

This fix targets modern browsers and aligns with our policy of supporting only up-to-date browser versions.